requirements.txt
absl-py==1.3.0
accelerate==0.14.0
aiohttp==3.8.3
aiosignal==1.2.0
albumentations==1.3.1
anyascii==0.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
cachetools==5.2.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
cloud-init==21.2
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.5
contractions==0.1.72
cryptography==2.8
cycler==0.11.0
cytoolz==0.12.1
datasets==2.13.1
dbus-python==1.2.16
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.7
distro==1.4.0
distro-info===0.23ubuntu1
dm-tree==0.1.8
entrypoints==0.3
etils==1.3.0
evaluate==0.4.0
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.11.0
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
gast==0.4.0
google-auth==2.14.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
grpcio==1.50.0
h5py==3.7.0
httplib2==0.14.0
huggingface-hub==0.13.4
humanfriendly==10.0
hyperlink==19.0.0
idna==2.8
imageio==2.31.1
importlib-metadata==5.0.0
importlib-resources==6.0.0
incremental==16.10.1
install==1.3.5
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
iterative-stratification==0.1.7
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.4.4
jupyter-console==6.4.4
jupyter-core==4.11.2
jupyter-server==1.21.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterthemes==0.20.0
kaggle==1.5.12
keras==2.13.1
keras-core==0.1.3
keras-cv==0.6.1
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.3
lesscpy==0.15.1
libclang==14.0.6
llvmlite==0.40.1
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
more-itertools==4.2.0
mpmath==1.2.1
multidict==6.0.2
multiprocess==0.70.14
namex==0.0.7
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==2.8.8
nltk==3.7
notebook==6.5.1
notebook-shim==0.2.0
numba==0.57.1
numpy==1.23.4
oauthlib==3.1.0
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.2.0
ply==3.11
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==4.23.4
psutil==5.9.3
ptsdae==1.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyahocorasick==1.4.4
pyarrow==12.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycocotools==2.0.6
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pypdf2==2.11.2
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-slugify==7.0.0
pytorch-lightning==1.7.7
pytz==2022.5
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
qudida==0.0.4
regex==2022.9.13
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
responses==0.18.0
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sniffio==1.3.0
sos==4.1
soupsieve==2.3.2.post1
ssh-import-id==5.10
stack-data==0.5.1
sympy==1.11.1
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.13.1
termcolor==2.1.0
terminado==0.17.0
text-unidecode==1.3
textsearch==0.0.24
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.9.2
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchmetrics==0.10.2
torchtext==0.14.0
torchvision==0.14.0+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.27.4
Twisted==18.9.0
typing-extensions==4.4.0
ubuntu-advantage-tools==27.2
ufw==0.36
ultralytics==8.0.141
unattended-upgrades==0.1
urllib3==2.0.4
visdom==0.2.4
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wordcloud==1.8.2.2
wrapt==1.14.1
xlrd==2.0.1
xxhash==3.2.0
yarl==1.8.1
zipp==3.16.2
zope.interface==4.7.1
